openai-custom/requirements.txt
pinecone-client
jupyter
langchain-pinecone
langchain-openai
langchain-text-splitters
langchain
python-dotenv

#langchain==0.1.1
#pinecone-client==2.2.4
#python-dotenv==1.0.0
#streamlit==1.29.0